+++ spec/requests/as_content_owner/video_spec.rb
@@ -422,6 +422,17 @@ describe Yt::Video, :partner do
         end
       end
 
+      describe 'views can be grouped by playback details' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+        let(:keys) { Yt::Collections::Reports::PLAYBACK_DETAILS.keys }
+
+        specify 'with the :by option set to :playback_details' do
+          views = video.views range.merge by: :playback_details
+          expect(views.keys - keys).to be_empty
+          expect(views.values).to all(be_an Integer)
+        end
+      end
+
       describe 'uniques can be retrieved for a single US state' do
         let(:state_code) { 'NY' }
         let(:result) { video.uniques since: date, by: by, in: location }